Added final models
This should be the final work on this notebook, for a while. I simply added a VotingClassifier as an ensemble method, combining the other models in a weighted fashion, giving more weight to the best classifier, based on previous cv work. The best performance achieved was from the SVM, at 83.58% out-of-sample performance, with relatively basic feature engineering.